moved platform authorization query and definition into a separate service under platform
