+++ src/domain/common/authorization-policy/authorization.policy.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthorizationPolicy } from './authorization.policy.entity';
 import { AuthorizationPolicyResolverFields } from './authorization.policy.resolver.fields';
-import { AuthorizationPolicyResolverQueries } from './authorization.policy.resolver.queries';
+
 import { AuthorizationPolicyService } from './authorization.policy.service';
 
 @Module({
@@ -11,11 +11,7 @@ import { AuthorizationPolicyService } from './authorization.policy.service';
     AuthorizationModule,
     TypeOrmModule.forFeature([AuthorizationPolicy]),
   ],
-  providers: [
-    AuthorizationPolicyService,
-    AuthorizationPolicyResolverFields,
-    AuthorizationPolicyResolverQueries,
-  ],
+  providers: [AuthorizationPolicyService, AuthorizationPolicyResolverFields],
   exports: [AuthorizationPolicyService],
 })
 export class AuthorizationPolicyModule {}
