+++ src/domain/common/authorization-policy/authorization.policy.service.ts
@@ -27,17 +27,13 @@ import { IAuthorizationPolicyRuleVerifiedCredential } from '@core/authorization/
 
 @Injectable()
 export class AuthorizationPolicyService {
-  private platformAuthorizationPolicy: IAuthorizationPolicy;
-
   constructor(
     @InjectRepository(AuthorizationPolicy)
     private authorizationPolicyRepository: Repository<AuthorizationPolicy>,
     private authorizationService: AuthorizationService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
     private readonly logger: LoggerService
-  ) {
-    this.platformAuthorizationPolicy = this.createPlatformAuthorizationPolicy();
-  }
+  ) {}
 
   reset(
     authorizationPolicy: IAuthorizationPolicy | undefined
@@ -166,15 +162,6 @@ export class AuthorizationPolicyService {
     return auth;
   }
 
-  inheritPlatformAuthorization(
-    childAuthorization: IAuthorizationPolicy | undefined
-  ): IAuthorizationPolicy {
-    return this.inheritParentAuthorization(
-      childAuthorization,
-      this.platformAuthorizationPolicy
-    );
-  }
-
   inheritParentAuthorization(
     childAuthorization: IAuthorizationPolicy | undefined,
     parentAuthorization: IAuthorizationPolicy | undefined
@@ -289,147 +276,5 @@ export class AuthorizationPolicyService {
     this.appendCredentialAuthorizationRules(authorization, newRules);
 
     return authorization;
-  }
-
-  private createPlatformAuthorizationPolicy(): IAuthorizationPolicy {
-    const platformAuthorization = new AuthorizationPolicy();
-
-    const credentialRules = this.createPlatformCredentialRules();
-
-    const platformAuthCredRules = this.appendCredentialAuthorizationRules(
-      platformAuthorization,
-      credentialRules
-    );
-
-    const privilegeRules = this.createPrivilegeRules();
-    return this.appendPrivilegeAuthorizationRules(
-      platformAuthCredRules,
-      privilegeRules
-    );
-  }
-
-  private createPlatformCredentialRules(): AuthorizationPolicyRuleCredential[] {
-    const credentialRules: AuthorizationPolicyRuleCredential[] = [];
-
-    const globalAdmin = new AuthorizationPolicyRuleCredential(
-      [
-        AuthorizationPrivilege.CREATE,
-        AuthorizationPrivilege.GRANT,
-        AuthorizationPrivilege.READ,
-        AuthorizationPrivilege.UPDATE,
-        AuthorizationPrivilege.DELETE,
-      ],
-      AuthorizationCredential.GLOBAL_ADMIN
-    );
-    credentialRules.push(globalAdmin);
-
-    const globalHubsAdmin = new AuthorizationPolicyRuleCredential(
-      [
-        AuthorizationPrivilege.CREATE,
-        AuthorizationPrivilege.GRANT,
-        AuthorizationPrivilege.READ,
-        AuthorizationPrivilege.UPDATE,
-        AuthorizationPrivilege.DELETE,
-      ],
-      AuthorizationCredential.GLOBAL_ADMIN_HUBS
-    );
-    credentialRules.push(globalHubsAdmin);
-
-    // Allow global admins to manage global privileges, access Platform mgmt
-    const globalAdminNotInherited = new AuthorizationPolicyRuleCredential(
-      [
-        AuthorizationPrivilege.GRANT_GLOBAL_ADMINS,
-        AuthorizationPrivilege.PLATFORM_ADMIN,
-        AuthorizationPrivilege.ADMIN,
-      ],
-      AuthorizationCredential.GLOBAL_ADMIN
-    );
-    globalAdminNotInherited.inheritable = false;
-    credentialRules.push(globalAdminNotInherited);
-
-    // Allow global admin Hubs to access Platform mgmt
-    const globalAdminHubsNotInherited = new AuthorizationPolicyRuleCredential(
-      [AuthorizationPrivilege.PLATFORM_ADMIN, AuthorizationPrivilege.ADMIN],
-      AuthorizationCredential.GLOBAL_ADMIN_HUBS
-    );
-    globalAdminHubsNotInherited.inheritable = false;
-    credentialRules.push(globalAdminHubsNotInherited);
-
-    // Allow global admin Communities to access Platform mgmt
-    const globalAdminCommunitiesNotInherited =
-      new AuthorizationPolicyRuleCredential(
-        [AuthorizationPrivilege.PLATFORM_ADMIN, AuthorizationPrivilege.ADMIN],
-        AuthorizationCredential.GLOBAL_ADMIN_COMMUNITY
-      );
-    globalAdminCommunitiesNotInherited.inheritable = false;
-    credentialRules.push(globalAdminCommunitiesNotInherited);
-
-    // Allow all registered users to query non-protected user information
-    const userNotInherited = new AuthorizationPolicyRuleCredential(
-      [AuthorizationPrivilege.READ_USERS],
-      AuthorizationCredential.GLOBAL_REGISTERED
-    );
-    userNotInherited.inheritable = false;
-    credentialRules.push(userNotInherited);
-
-    // Allow hub admins to create new organizations
-    const hubAdminsNotInherited = new AuthorizationPolicyRuleCredential(
-      [
-        AuthorizationPrivilege.CREATE_ORGANIZATION,
-        AuthorizationPrivilege.ADMIN,
-      ],
-      AuthorizationCredential.HUB_ADMIN
-    );
-    hubAdminsNotInherited.inheritable = false;
-    credentialRules.push(hubAdminsNotInherited);
-
-    // Allow challenge admins to create new organizations + access platform admin
-    const challengeAdminsNotInherited = new AuthorizationPolicyRuleCredential(
-      [
-        AuthorizationPrivilege.CREATE_ORGANIZATION,
-        AuthorizationPrivilege.ADMIN,
-      ],
-      AuthorizationCredential.CHALLENGE_ADMIN
-    );
-    challengeAdminsNotInherited.inheritable = false;
-    credentialRules.push(challengeAdminsNotInherited);
-
-    // Allow Opportunity admins to access admin
-    const opportunityAdminNotInherited = new AuthorizationPolicyRuleCredential(
-      [AuthorizationPrivilege.ADMIN],
-      AuthorizationCredential.OPPORTUNITY_ADMIN
-    );
-    opportunityAdminNotInherited.inheritable = false;
-    credentialRules.push(opportunityAdminNotInherited);
-
-    // Allow Organization admins to access platform admin
-    const organizationAdminsNotInherited =
-      new AuthorizationPolicyRuleCredential(
-        [AuthorizationPrivilege.ADMIN],
-        AuthorizationCredential.ORGANIZATION_ADMIN
-      );
-    organizationAdminsNotInherited.inheritable = false;
-    credentialRules.push(organizationAdminsNotInherited);
-
-    return credentialRules;
-  }
-
-  private createPrivilegeRules(): AuthorizationPolicyRulePrivilege[] {
-    const privilegeRules: AuthorizationPolicyRulePrivilege[] = [];
-
-    const createPrivilege = new AuthorizationPolicyRulePrivilege(
-      [
-        AuthorizationPrivilege.CREATE_HUB,
-        AuthorizationPrivilege.CREATE_ORGANIZATION,
-      ],
-      AuthorizationPrivilege.CREATE
-    );
-    privilegeRules.push(createPrivilege);
-
-    return privilegeRules;
-  }
-
-  getPlatformAuthorizationPolicy(): IAuthorizationPolicy {
-    return this.platformAuthorizationPolicy;
   }
 }
