+++ src/domain/community/user/user.resolver.queries.ts
@@ -6,7 +6,6 @@ import { GraphqlGuard } from '@core/authorization';
 import { AuthorizationService } from '@core/authorization/authorization.service';
 import { AgentService } from '@domain/agent/agent/agent.service';
 import { CredentialMetadataOutput } from '@domain/agent/verified-credential/dto/verified.credential.dto.metadata';
-import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { UUID_NAMEID_EMAIL } from '@domain/common/scalars';
 import { UseGuards } from '@nestjs/common';
 import { Args, Float, Query, Resolver } from '@nestjs/graphql';
@@ -16,12 +15,13 @@ import { PaginationArgs, PaginatedUsers } from '@core/pagination';
 import { UserService } from './user.service';
 import { IUser } from './';
 import { UserFilterInput } from '@core/filtering';
+import { PlatformAuthorizationService } from '@src/platform/authorization/platform.authorization.service';
 
 @Resolver(() => IUser)
 export class UserResolverQueries {
   constructor(
     private authorizationService: AuthorizationService,
-    private authorizationPolicyService: AuthorizationPolicyService,
+    private platformAuthorizationService: PlatformAuthorizationService,
     private userService: UserService,
     private agentService: AgentService
   ) {}
@@ -53,7 +53,7 @@ export class UserResolverQueries {
   ): Promise<IUser[]> {
     await this.authorizationService.grantAccessOrFail(
       agentInfo,
-      this.authorizationPolicyService.getPlatformAuthorizationPolicy(),
+      this.platformAuthorizationService.getPlatformAuthorizationPolicy(),
       AuthorizationPrivilege.READ_USERS,
       `users query: ${agentInfo.email}`
     );
@@ -73,7 +73,7 @@ export class UserResolverQueries {
   ): Promise<PaginatedUsers> {
     await this.authorizationService.grantAccessOrFail(
       agentInfo,
-      this.authorizationPolicyService.getPlatformAuthorizationPolicy(),
+      this.platformAuthorizationService.getPlatformAuthorizationPolicy(),
       AuthorizationPrivilege.READ_USERS,
       `users query: ${agentInfo.email}`
     );
@@ -93,7 +93,7 @@ export class UserResolverQueries {
   ): Promise<IUser> {
     await this.authorizationService.grantAccessOrFail(
       agentInfo,
-      this.authorizationPolicyService.getPlatformAuthorizationPolicy(),
+      this.platformAuthorizationService.getPlatformAuthorizationPolicy(),
       AuthorizationPrivilege.READ_USERS,
       `user query: ${agentInfo.email}`
     );
@@ -112,7 +112,7 @@ export class UserResolverQueries {
   ): Promise<IUser[]> {
     await this.authorizationService.grantAccessOrFail(
       agentInfo,
-      this.authorizationPolicyService.getPlatformAuthorizationPolicy(),
+      this.platformAuthorizationService.getPlatformAuthorizationPolicy(),
       AuthorizationPrivilege.READ_USERS,
       `users query: ${agentInfo.email}`
     );
